+++ R/ml_fit.R
@@ -4,13 +4,14 @@
 #' aggregate controls using an algorithm given as parameter.
 #'
 #' @usage
-#' ml_fit(algorithm, ref_sample, controls, field_names, ...)
+#' ml_fit(algorithm, ref_sample, controls, field_names, verbose, ...)
 #' @param algorithm Algorithm to use
 #' @param ref_sample Reference sample.  Alternatively, a configuration as
 #'   returned by \code{\link{import_IPAF_results}}; in this case, the other
 #'   arguments are ignored.
 #' @param controls Control totals
 #' @param field_names Names of special columns in the data
+#' @param verbose If \code{TRUE}, print diagnostic output.
 #' @param ... Further parameters passed to the algorithm
 #' @return An object of class \code{ml_fit}, essentially a named list.
 #' @export
@@ -18,14 +19,17 @@
 #' path <- system.file("extdata/minitoy", package="MultiLevelIPF")
 #' ml_fit(algorithm = "entropy_o", ref_sample = import_IPAF_results(path))
 ml_fit <- function(algorithm = c("entropy_o"),
-                   ref_sample, controls, field_names, ...) {
+                   ref_sample, controls, field_names, verbose = FALSE, ...) {
   algorithm <- match.arg(algorithm)
   fun.name <- sprintf("ml_fit_%s", algorithm)
   if (!exists(fun.name))
     stop("Unknown algorithm:", algorithm)
 
   get(fun.name)(
-    ref_sample=ref_sample, controls=controls, field_names=field_names, ...)
+    ref_sample = ref_sample,
+    controls = controls,
+    field_names = field_names,
+    verbose = verbose, ...)
 }
 
 .patch_ml_fit_args <- function() {